apps/scorecard/requirements.txt
streamlit>=1.36
pandas>=2.2
matplotlib>=3.8
python-dotenv>=1.0
azure-cosmos>=4.7
